requirements.txt
rich
streamlit
pandas
yara-python
pefile
python-dotenv
requests
fpdf2
